Close #11. Move signInJwt and signUpJwt methods to new AuthRestController.
